+++ packages/thumbprint-react/components/Wrap/index.tsx
@@ -1,31 +1,24 @@
 import React from 'react';
-import PropTypes from 'prop-types';
 import classNames from 'classnames';
 import styles from './index.module.scss';
 
-const propTypes = {
+interface PropTypes {
     /**
      * Component to be wrapped.
      */
-    children: PropTypes.node,
+    children?: React.ReactNode;
     /**
      * Breakpoint at which the wrapper should remove horizontal margin to bleed to the edge of the
      * viewport. Defaults to `undefined`, in which case it never does this.
      */
-    bleedBelow: PropTypes.oneOf(['small', 'medium', 'large']),
+    bleedBelow?: 'small' | 'medium' | 'large';
     /**
      * A selector hook into the React component for use in automated testing environments.
      */
-    dataTest: PropTypes.string,
-};
+    dataTest?: string;
+}
 
-const defaultProps = {
-    children: undefined,
-    bleedBelow: undefined,
-    dataTest: undefined,
-};
-
-export default function Wrap({ children, bleedBelow, dataTest }) {
+export default function Wrap({ children, bleedBelow, dataTest }: PropTypes): JSX.Element {
     return (
         <div
             className={classNames({
@@ -40,6 +33,3 @@ export default function Wrap({ children, bleedBelow, dataTest }) {
         </div>
     );
 }
-
-Wrap.defaultProps = defaultProps;
-Wrap.propTypes = propTypes;