+++ packages/thumbprint-react/components/Image/get-scroll-parent.ts
@@ -4,7 +4,7 @@
  * It is adapted from:
  * https://stackoverflow.com/a/42543908/316602
  */
-export default function getScrollParent(element) {
+export default function getScrollParent(element: Element): Element {
     let style = getComputedStyle(element);
     const excludeStaticParent = style.position === 'absolute';
     const overflowRegex = /(auto|scroll)/;
@@ -12,7 +12,7 @@ export default function getScrollParent(element) {
     if (style.position === 'fixed') return document.body;
 
     // eslint-disable-next-line no-cond-assign
-    for (let parent = element; (parent = parent.parentElement); ) {
+    for (let parent: Element | null = element; (parent = parent.parentElement); ) {
         style = getComputedStyle(parent);
         if (excludeStaticParent && style.position === 'static') {
             // eslint-disable-next-line no-continue