+++ packages/thumbprint-react/index.ts
@@ -46,7 +46,7 @@ export {
     ServiceCardTitle,
     ServiceCardImage,
 } from './components/ServiceCard/index.jsx';
-export { default as StarRating } from './components/StarRating/index.jsx';
+export { default as StarRating } from './components/StarRating/index';
 export { default as Textarea } from './components/Textarea/index.jsx';
 export { default as Tooltip } from './components/Tooltip/index';
 export { Title, Text } from './components/Type/index';